+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Added
+-----
+- Send custom signal ``post_duplicate`` when objects are duplicated since
+  regular Django signals are not sent
+
+
 ===================
 30.1.0 - 2022-04-15
 ===================